rename @contexts to @globals, move schemas out to root and alias @schemas, remove Utils append from dateTimeUtils
