+++ app/components/Forms/TaskForm.tsx
@@ -13,9 +13,9 @@ import { v4 as uuidv4 } from 'uuid';
 import dayjs, { Dayjs } from 'dayjs';
 import { today, getLocalTimeZone } from '@internationalized/date';
 // types
-import { Task, Label } from '@defines/schemas';
+import { Task, Label } from '@schemas';
 // utils
-import { dateToCalendarDate } from '@utils/datetimeUtils';
+import { dateToCalendarDate } from '@utils/datetime';
 
 type Props = {
   openColumn: string;
