+++ app/globals/WorkspacesContext.tsx
@@ -6,7 +6,7 @@ import { v4 as uuidv4 } from 'uuid';
 // date
 import { CalendarDate } from '@internationalized/date';
 // types
-import { Workspace, Project, Sprint, Column, Task } from '@defines/schemas';
+import { Workspace, Project, Sprint, Column, Task } from '@schemas';
 // sample data
 import sampleWorkspaces from './sampleData';
