+++ app/globals/sampleData.ts
@@ -1,7 +1,7 @@
 import { genWorkspace } from './sampleDataHelper';
 import { v4 as uuidv4 } from 'uuid';
 // types
-import { Workspace } from '@defines/schemas';
+import { Workspace } from '@schemas';
 
 // generate workspaces
 const sampleWorkspaceTitles: string[] = [
